+++ lib/src/piece/text.dart
@@ -59,18 +59,6 @@ sealed class TextPiece extends Piece {
     }
   }
 
-  /// Sets [selectionStart] to be [start] code units after the end of the
-  /// current text in this piece.
-  void startSelection(int start) {
-    _selectionStart = _adjustSelection(start);
-  }
-
-  /// Sets [selectionEnd] to be [end] code units after the end of the
-  /// current text in this piece.
-  void endSelection(int end) {
-    _selectionEnd = _adjustSelection(end);
-  }
-
   /// Adjust [offset] by the current length of this [TextPiece].
   int _adjustSelection(int offset) {
     for (var line in _lines) {
@@ -163,7 +151,7 @@ class CommentPiece extends TextPiece {
   /// Whitespace at the end of the comment.
   final Whitespace _trailingWhitespace;
 
-  CommentPiece([this._trailingWhitespace = Whitespace.none]);
+  CommentPiece(this._trailingWhitespace);
 
   @override
   void format(CodeWriter writer, State state) {
@@ -178,6 +166,32 @@ class CommentPiece extends TextPiece {
 
   @override
   void forEachChild(void Function(Piece piece) callback) {}
+}
+
+/// A piece for the special `// dart format off` and `// dart format on`
+/// comments that are used to opt a region of code out of being formatted.
+class EnableFormattingCommentPiece extends CommentPiece {
+  /// Whether this comment disables formatting (`format off`) or re-enables it
+  /// (`format on`).
+  final bool _enabled;
+
+  /// The number of code points from the beginning of the unformatted source
+  /// where the unformatted code should begin or end.
+  ///
+  /// If this piece is for `// dart format off`, then the offset is just past
+  /// the `off`. If this piece is for `// dart format on`, it points to just
+  /// before `//`.
+  final int _sourceOffset;
+
+  EnableFormattingCommentPiece(this._sourceOffset, super._trailingWhitespace,
+      {required bool enable})
+      : _enabled = enable;
+
+  @override
+  void format(CodeWriter writer, State state) {
+    super.format(writer, state);
+    writer.setFormattingEnabled(_enabled, _sourceOffset);
+  }
 }
 
 /// A piece that writes a single space.