+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 * Allow passing a language version to `DartFomatter()`. Formatted code will be
   parsed at that version. If omitted, defaults to the latest version. In a
   future release, this parameter will become required.
+* Allow opting out of formatting for a region of code using `// dart format off`
+  and `// dart format on` comments. Note: This only works using the new tall
+  style and requires passing the `--enable-experiment=tall-style` experiment
+  flag (#361).
 * Preserve type parameters on old-style function-typed formals that also use
   `this.` or `super.` (#1321).
 * Remove temporary work around for analyzer 6.2.0 from dart_style 2.3.6.